STEP 14. Move out asset detail nav.
